+++ src/utils/createGenerateClassName.js
@@ -1,11 +1,17 @@
 /* @flow */
+import warning from 'warning'
 import type {Rule, generateClassName} from '../types'
+import StyleSheet from '../StyleSheet'
 
 const globalRef = typeof window === 'undefined' ? global : window
-const namespace = '__JSS_VERSION_COUNTER__'
-if (globalRef[namespace] == null) globalRef[namespace] = 0
+const ns = '2f1acc6c3a606b082e5eef5e54414ffb'
+if (globalRef[ns] == null) globalRef[ns] = 0
 // In case we have more than one JSS version.
-const jssCounter = globalRef[namespace]++
+const jssCounter = globalRef[ns]++
+
+const maxRules = 1e10
+
+const env = process.env.NODE_ENV
 
 /**
  * Returns a function which generates unique class names based on counters.
@@ -15,5 +21,23 @@ const jssCounter = globalRef[namespace]++
 export default (): generateClassName => {
   let ruleCounter = 0
 
-  return (rule: Rule): string => `${rule.key}-${jssCounter}-${ruleCounter++}`
+  return (rule: Rule, sheet?: StyleSheet): string => {
+    ruleCounter += 1
+
+    if (ruleCounter > maxRules) {
+      warning(
+        false,
+        'You might have a memory leak. Rule counter is at %s.',
+        ruleCounter
+      )
+    }
+
+    if (env === 'production') {
+      return `c${jssCounter}${ruleCounter}`
+    }
+
+    const prefix = sheet ? (sheet.options.classNamePrefix || '') : ''
+
+    return `${prefix}${rule.key}-${jssCounter}-${ruleCounter}`
+  }
 }