+++ src/utils/isDynamicValue.js
@@ -1,0 +1,3 @@
+import isObservable from 'is-observable'
+
+export default value => typeof value === 'function' || isObservable(value)